add more unit tests to managers
In this commit we add several unit tests that covers some managers.
We chose to write unit tests for managers that are closest to Synergy
core functionalities in a first stage.
Also, we test only method specific to each manager as the base methods
of Manager class have been tested in Synergy.

Managers with unit tests:
- FairShareManager
- QueueManager
- QuotaManager (and DynamicQuota)
- SchedulerManager

Change-Id: I831649e48b36bbaf94f6790fbf1a8954636fac32