Remove versions for required packages
Requirement versions were set in the first place to limit the
buginess. However, after some time it appears that it made more
harm than good, as discussed in:
https://review.openstack.org/#/c/388003/

In this commit we remove the versions for the requirements, except
for the following:
- Nova
- synergy-service

In the future, if a specific dependency version is causing some issue
we should reintroduce the version for this dependency to block the
unstable version.

Sem-Ver: bugfix
Change-Id: If65e670a18a495a5c3f5c08c80b6b1d3abec5747